+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 ## Beta Versions
 
+### v0.6.1
+- Better my account section on Settings
+- Add Your Data link on username dropdown
+- Add a quote from Kobe Bryant, RIP
+
 ### v0.5.14
 - Fix Linux build toolbar bug
 
 ## Pre-Beta Versions
 
-No changelog for pre-beta versions
+No changelog was recorded for pre-beta versions